Reverted type decof::integer. Fixed testcase.

+++ include/decof/types.h
@@ -17,6 +17,7 @@
 #ifndef DECOF_TYPES_H
 #define DECOF_TYPES_H
 
+#include <cctype>
 #include <string>
 #include <vector>
 
@@ -25,7 +26,7 @@ namespace decof
 
 // Scalar parameter types
 typedef bool boolean;
-typedef std::intmax_t integer;
+typedef std::int32_t integer;
 typedef double real;
 typedef std::string string;
 
@@ -42,7 +43,7 @@ struct sequence : public std::vector<T>
 };
 
 typedef sequence<bool> boolean_seq;
-typedef sequence<std::intmax_t> integer_seq;
+typedef sequence<std::int32_t> integer_seq;
 typedef sequence<double> real_seq;
 typedef sequence<std::string> string_seq;
 typedef sequence<binary> binary_seq;
